Move view/serializers into dedicated folders
