+++ src/researchhub/serializers/__init__.py
@@ -1,0 +1,2 @@
+from .asset_upload_serializer import AssetUploadSerializer
+from .serializers import DynamicModelFieldSerializer
