+++ src/researchhub/views/__init__.py
@@ -1,0 +1,2 @@
+from .asset_upload_view import AssetUploadView
+from .views import *
